requirements1.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sentence-transformers==2.2.2
torch==2.1.0
duckduckgo-search==3.9.6
httpx==0.25.1
python-dateutil==2.8.2
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
redis==5.0.1
transformers==4.35.2
feedparser
transformers==4.35.2
torch==2.1.0
sentencepiece==0.1.99
sacremoses==0.1.1
langdetect==1.0.9